pkg/supervisor: Split so each datatype has its own file [ci-skip]
